requirements.txt
blinker==1.9.0
click==8.3.0
dominate==2.9.1
Flask==3.1.2
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
python-dotenv==1.1.1
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
